SHOW: change date and time data row representation
The current code only prints seconds in date and time columns
(represented by 'T'). Sometimes millisecond precision is important,
hence, let's append it. Note that it can break applications that does
not expect the millisecond component. However, if a code does not parse
the whole date and time accordingly it is already broken and should be
fixed. The 'T' format type is rarely used (only 2 columns in all SHOW
commands) so if this change breaks applications, the damage will be
marginal.