Add kill_client command (#1147)
This adds a kill_client command to the admin console. This command is
the equivalent of the Postgres pg_terminate_backend() function. Similar
functionality also exists in other Postgres connection poolers.

In general this functionality is a useful last resort to undisciplined
users opening a lot of connections and which block out other users. This
is especially prominent with front end GUI DB applications that open a
new session for every SQL editor tab.