Add identifier to socket (#1172)
Adds a numeric identifier to PgSocket. The main reason is to use it as an unique
identifier for KILL_CLIENT command (commit 1dbde965e67). The current identifier
is the PgSocket pointer but due to the reuse nature of slab allocation, there
might cancel another client (if the client connection ends between SHOW CLIENTS
and KILL_CLIENT). Let's not wait for complaints as we discussed in #1147 and
closes the gap now.