Add client user connection limit (#1137)
This PR adds a user level client connection limit. I believe that
currently user connections can only be limited by how many people are
given connections to the database backends but there is no way to limit
the number of connections to pgbouncer itself. Combined with the global
connection limit this means that one user could use up all connections
on the pgbouncer instance. This PR proposes a fix to this problem.

Related to #1136 (because it partially fixes it)