Attempt to fix inconsistent error (#1152)
This changes `server_login_retry` to also cache the error message. This
makes it much easier to debug an authentication problem between
PgBouncer and Postgres, because now the client always sees the message,
not just the first time. Before you'd have to check the PgBouncer logs
to find the actual error message.

Fixes #1149