Handle auth_query correctly for replication connections (#1166)
The way we send an auth_query is a bit strange. We temporarily set the
pool of the client to the pool for the auth_user, and then continue
connection startup like normal. This broke for replication connections
because we would send the replication parameter for that connection too.
This fixes that by only doing our special connection setup logic for
replication connections if we're not currently trying to send an
auth_query.

Fixes #1164