Do not specify the hostname in tests with ipv6 local host
This is a very unsatisfying fix. I found that when the IPV6 localhost
postgres instance needs its config changed in certain specs to bind to ::1
that does not appear to work at all when -h is passed in the
command-line.

This sort of makes sense. Maybe just on "restarts" we should not specify
host but this works for now and at least _appears_ intentional to me.

Open to tweaking this however, though I will admit that longer-term using
the IP address to distinguish "this instance" from "that" won't work at
all in an IPv6 world so perhaps everything should bind to 127.0.0.1 on
differing ports instead of this?