Set Basic Constraints X.509 v3 extension (#1219)
It appears that recent versions of the OpenSSL library require the
"Basic Constraints" X509 extension to be properly configured for CA
certificates. The x509v3_config(5) man page states:
"A CA certificate must include the basicConstraints name with the CA
parameter set to TRUE. An end-user certificate must either have CA:FALSE
or omit the extension entirely."
Currently, the basic constraints are not set when generating CA
certificates and SSL tests fail on distros with new OpenSSL libraries
such as Rocky Linux 9. This PR fixes the issue.

For reference:
https://docs.openssl.org/3.4/man5/x509v3_config/#basic-constraints

Fixes #1210 
Fixes #1031