Fix Client Connection Count Decrementing (#1226)
Fixes [#1225](https://github.com/pgbouncer/pgbouncer/issues/1225). 

It seems that sometimes when a connection is closed pgbouncer does not
make a call to `disconnect_client`. Instead it goes through
`disconnect_client_sqlstate`. This PR moves the count decrementing from
`disconnect_client` and to `disconnect_client_sqlstate` so that it gets
counted more often.

Added a test case that proves this out which fails when running against
the current state of pgbouncer.