option to disable default pool size (#1227)
This PR implements the non default option of choosing zero as the
default pool size and not having the restriction apply. This is useful
in situations where you would rather limit connections by user or
database instead of pool. Without this option it can be easy to forget
to make sure that default pool size is larger than the limit you
actually want to apply.

Co-authored-by: Andrew Jackson <andrewjackson2988@gmail.com>