Skip tests for Close messages on Debian oldstable
Although Debian proper does not release PG/libpq=17 for bullseye
(oldstable) the PostgreSQL Debian team does release those packages for
bullseye via "pgdq".

Unfortunately, `psycopg` adds `Close` support in version 3.2.0 and it
would be a much _bigger_ lift to support that package.

This leaves a tiny gap in our integration testing for bullseye but I
think we can be confident that pgdg users will be able to
`close_prepared`.

Debian Trixie becomes Debian `stable` in a few months, and so this will
be moot for OSS packagers when bullseye moves to `oldoldstable` but I
and other users have a vested interest in being able to use the latest
pgbouncer while migrating to Debian stable.