Account for VALID UNTIL in auth_query (fixes CVE-2025-2291)
Previously PgBouncer did not take into account the VALID UNTIL of a user
password when querying for password hashes using its auth_query. So if
PgBouncer is used as a transparent proxy in front of Postgres it could
allow passwords that had already expired.

To solve this issue this changes the default auth_query and the examples
of custom auth_query functions in the documentation to take VALID UNTIL
into account.

Since this can be considered a security issue in setups where VALID
UNTIL is used to limit exposure of leaked passwords, this is tracked as
CVE-2025-2291.