Use devel suffix for development
The current version during development is the latest release version.
That's not ideal because it confuses the stable release with a
development code.

The new policy states that for a release, we remove the "devel" suffix
and the first commit after release is to (a) increment the major and/or
minor version and (b) add the "devel" suffix back. It means if we
release 1.24.0, the next commit after release should update version from
1.24.0 to 1.25devel.