Also check the second digit in the openssl version when creating certificates (#1254)
PR #1221 introduced new changes needed to ensure that on
systems with old openssl (v1.0.x) certificates are generated correctly
by skipping the -addext option. But I found out that on the
astra-leningrad-8.1 system, on which the openssl 1.1.f version, the
check proposed in the above-mentioned commit will not work. This
adds a check for the second digit in the version number.