Match the MAX_PASSWORD value in bouncer.h (#1251)
Doing a bit of testing in AWS I discovered that the token (which is over
1024 bytes) would cause pg_bouncer to segfault when the server requests
SASL auth. With some digging I found that the MAX_PASSWORD_LENGTH in
`saslprep.c` didn't match what was defined in bouncer.h.

This small patch just ensures the value matches. There is probably a
better way to have this defined in one place instead of two. Alas this
is a quick fix for those who may run into this issue.