Add error message on reload fail (#1231)
This PR adds an error message to the reload command on the admin console
if there is a failure. Currently the only way to know if the admin
command worked is to have the logs open while you enter `RELOAD` on the
admin screen. If you do not you might miss the fact that your
configuration change failed. The purpose of this PR is to make the
failed reload more obvious. Even if the entire message will not be
displayed in the admin console it will at least tell the user that
something went wrong so they can look at the logs.