Add LDAP authentication to pgbouncer (#731)
Previously the only way to support ldap is by using PAM in pgbouncer.
Configuring pam with ldap is quite hard. So this commit adds ldap
support to allow more users authenticate with ldap in pgbouncer. Because
some of the ldap library functions are run in synchronous way, we follow
the threading model that's also used for PAM.

The configuration uses the same style as Postgres, i.e. it's possible to
configure LDAP through the hba file. In addition it's also possible to
use a few newly introduced config keys to configure LDAP globally (in
setups that don't use HBA).

The implementation is also heavily inspired by/copied from Postgres.