Add transaction timeout (#1242)
This PR adds a transaction timeout global setting to pgbouncer. In the
current state there is an idle_in_transaction_timeout and a
query_timeout but no transaction_timeout. This is meant to combat issues
where a user starts a transaction and then proceeds to wait and issue
`SELECT 1;` repeatedly thus making the transaction not idle. There is an
equivalent setting on the postgres side but it is easily overridden by
any user.

Not looking to get this in before the next release but I believe that
this functionality could be very useful and just want to put a PR out
here for visibility.