Queue notification (#1264)
This PR adds a configuration item `client_queue_notify_seconds`. By
default this is disabled but when set it will alert a client that is
awaiting a server connection after they have been queued for the amount
of time in `client_queue_notify_seconds` with the message "No server
connection available in postgres backend, client being queued".

As is it can be confusing for users that are queued as the terminal just
looks blank and appears to hang. This PR allows pgbouncer to provide
some visual feedback to the client as to why their query hasn't returned
instantly.