Enable Admin Console Access During Safe Shutdown (#1305)
In it's current implementation, during safe shutdowns there is no way to
access the admin console. This is problematic if you want to log in, see
who is on, alert tenants to restart their apps, etc. Currently the only
way to gain visibility about who is connected to your app is via `lsof`
and other cli tools but even here you lose visibility as to what
database they are connected to, what user, etc.

This PR attempts to provide a way for admin users to continue to access
the admin console by keeping the Unix Socket open during the safe
shutdown period.

---------

Co-authored-by: Andrew Jackson <andrewjackson2988@gmail.com>