Test 'test_server_id' is updated to support the slow machines
When this test works on the slow machines or we run this test under debugger, PgBouncer may creates TWO connections instead ONE and this test is starting to fail.

This needs to be taken into account.

Moreover, we will make timeout explicitly now and detect the situation when PgBouncer does not create two connection.

It is an indicator of the slow machine, too.

---
When we detect a slow machine we will send warning message.

---
See #1283.