Handle test flakiness in test_servers_disconnect_when_changing_tls_config (#1322)
The problem in the test that's solved here as an example was that half a
second wasn't always enough for PgBouncer to close the connection as
expected. This adds a wait_until method, to allow for easy polling until
the required condition is true.

This does a similar fix to #1320, but by introducing this new method
similar test failures can be fixed with a lot less code.

Reported in #1283
Closes #1320