Remove uthash submodule
The goal is to remove all uses of git submodules, because they are
inconvenient for developers and users.  Instead, keep a local copy of
the files.

This copies over what we need from the uthash submodule into our own
git tree.  We only need the uthash.h file.  We place it into
include/common/, because uthash_lowercase.h is already there, and
include/common/ is excluded from uncrustify.