Remove lib submodule
The goal is to remove all uses of git submodules, because they are
inconvenient for developers and users.  Instead, keep a local copy of
the files.

This copies over all the files from libusual in their current state.
Some of them are now redundant (e.g., .cirrus.yml), but cleaning that
up is left as a separate step.

The libusual test suite is currently not wired into .cirrus.yml, but
you can run it manually by running configure; make; make check etc. as
before in the lib directory.  Improving this will also be a future
project.

We will from now on maintain the required parts of libusual inside the
pgbouncer source tree and abandon the separate libusual/libusual
repository.