make dist uses git archive
This changes "make dist" to directly use "git archive", rather than
the custom shell script it currently runs.

This is to make the creation of the distribution tarball more
reliable, as this does not require manual maintenance of various
*_DIST make variables.  Eventually, it will also make the distribution
tarball more directly traceable to the git repository.  For the time
being, we still include files produced by autogen.sh in the tarball,
for ease of use.

The techniques used are the same as in postgres/postgres@619bc23a1a2.

(This required getting rid of git submodules, since git-archive does
not support submodules.)