Fix include order hygiene
Commit 8765458 added another #include before "system.h" in bouncer.h.
This is not good, because we need system.h and ultimately config.h to
be included before all system header files.

A particular outcome is that the getaddrinfo_a()-based adns not longer
compiled, because that requires _GNU_SOURCE to be defined before
including system headers.

The #include was added to get SCRAM_KEY_LEN, but as can be seen in
nearby field definitions, we already chose to hardcode 32 instead of
using the symbol, for this and related reasons.  So do this here, too.