Add support to set ciphers for TLSv1.3 (#1352)
For TLS v1.3 we need to use a different API to set the ciphers on
OpenSSL. This adds new config options to configure these ciphers. The
new options are called `client_tls13_ciphers` and
`server_tls13_ciphers`. These names aligns well with Postgres its new
`ssl_tls13_ciphers` option, which was introduced in PG18.

This PR was previously open on libusual:
https://github.com/libusual/libusual/pull/72

---------

Co-authored-by: Jelte Fennema-Nio <postgres@jeltef.nl>