Update the existing and new files
After copying (over) the required files, a few changes are required to
compile the code. The challenge was to apply minimal changes to make it
work.  system.h and postgres_compat.h does the trick plus changing some
references to header files that is now in the include/common directory.
Makefile included the new required files.