ci: Add more configure args to Red Hat and Alpine jobs
The idea is to have these jobs test more features and have the
configuration be similar to what the downstream packagers are using,
so it's more realistic overall.

This adds --with-pam to the Red Hat jobs, and --with-cares to RHEL
>=9.  Also --with-cares to Alpine.

The FreeBSD and macOS jobs appear to be in good shape in this respect
already (no c-ares in either case in the downstream packages).