Update the vendored PG files for PgBouncer (#1356)
After copying (over) the required files, a few changes are required to
be able to compile the code within the constraints of PgBouncer source
code and build system. This mostly requires including `postgres_compat.h`.

Apart from that it also removes any references to MD5 and SHA1 from the
cryto files, since we don't need those for SCRAM (which is SHA256 only)

The full commit DOES NOT COMPILE yet, because the PgBouncer files have
not been updated to use the new PG APIs. This will be fixed in the next
commit.