Fix LDAP tests on FreeBSD
The script start_openldap_server.sh hardcoded /bin/bash, which does
not exist on FreeBSD.  The script works just fine with /bin/sh, so use
that instead.

Also, generalize the search for various openldap-related paths to
handle the FreeBSD installation location.

With that fixed, we can enable LDAP on the CI FreeBSD task.  We need
to install openldap client and server packages for that.