Enable LDAP tests on macOS
We need to look for the LDAP files in various locations.  I copied the
locations used by PostgreSQL for its LDAP tests; this supports both
Homebrew and MacPorts locations.  (The macOS native LDAP libraries are
not supported by neither PgBouncer nor PostgreSQL.)  Then we can
un-skip the tests on macOS and enable them in CI.