Avoid rare double call to disconnect_server for cancellations
It was possible for disconnect_server to be called twice on a server
that was used to forward a cancel request. This would happen when
`SEND_CancelRequest` would fail. If that happened we would have already
linked the client and server, so by calling `disconnect_client` it would

This simplifies the code by linking the server and client earlier in
`forward_cancel_request`. This way we can relying on the
`disconnect_client` calls in `forward_cancel_request` to also always
disconnect the server, so we don't need the additional
`disconnect_server` call in `handle_connect` anymore.

For easier understanding this also move the server state change that
happens on success into that `forward_cancel_request`, so that all the
related code is together.

Since the `disconnect_client` code can also close servers that are in
the `SV_LOGIN` state, we also set `server->ready` for those cases. To
avoid tracking this as a connection failure, like we did in
`handle_connect` before.