Rename ldap_options to auth_ldap_options
The argument in commit 1a95d51 that the "auth" prefix is unnecessary
is flawed.  There are also other ways in which LDAP could be used in a
PostgreSQL stack, and they would use different LDAP options.  Explicit
is better.  Also, PgBouncer groups all related settings under common
prefixes, and they sort close together, so it seems good to maintain
that.