Fix localtime_r() on MinGW
The libusual substitute implementation of localtime_r() was faulty.
It populated the tm_isdst field with garbage.  This was apparently not
noticed, but it would be noticed if you used strftime() with %Z,
because then it would always show a "Daylight" time zone even if it
was not daylight-saving time.

To fix, adopt the solution used by PostgreSQL code: We can use the
localtime_r() provided by MinGW if we define _POSIX_C_SOURCE.  (If we
ever support MSVC, we can also adapt a solution for that from
PostgreSQL code, but that is for another day.)  The implementation in
libusual is deleted.