Fix errors with ad-hoc SCRAM auth after reconnect to server
An unfortunate bug got introduced with the new caching of ad-hoc SCRAM
secrets (#1338). On any reconnections to the server the ad-hoc scram
secret would be treated as a pass-through scram secret (even though we
know for sure it's different because it was generated by pgbouncer).
This would cause any subsequent connection attempts to the server to
fail with "password authentication failed".

This fixes that issue and adds a regression test.

Related to #1338
Fixes #1418