Make warnings in tests behave as errors
In PR #1264 a few tests were not correctly awaiting all coroutines. This
caused warnings in the tests (and I think also some flakiness). This
fixes that by sprinkling await in a few more places. To avoid this
happening again in the future this also starts treating warnings in
tests as errors. If later we want to ignore some warnings we can
explicitly do that.