Fix errors with ad-hoc SCRAM auth after reconnect to server (#1432)
An unfortunate bug got introduced with the new caching of ad-hoc SCRAM
secrets (#1338). On any reconnections to the server the ad-hoc scram
secret would be treated as a pass-through scram secret (even though we
know for sure it's different because it was generated by pgbouncer).
This would cause any subsequent connection attempts to the server to
fail with "password authentication failed".

This fixes that issue by correctly marking the credentials as adhoc. It
also adds a regression test.

In passing it also starts sending the pgbouncer password on admin
connections in our tests, because that's needed to make the new test
pass on Windows (which doesn't have UNIX sockets).

Related to #1338
Fixes #1418