Correctly handle large and partial server login packets
This issue was discovered when there was an intermediate proxy like so:
pgbouncer -> other proxy -> upstream db. If for whatever reason `other
proxy` split a login packet into multiple parts, we would bail the
connection. This fix is modeled after the fix in 
https://github.com/pgbouncer/pgbouncer/pull/1058.

If the packet is simply incomplete we yield the sbuf back and wait for
the next event loop tick. If it's bigger than the sbuf_len it uses the
existing packet buffering logic.